refactor: 전진조건 값 구성 수정
