+++ src/main/java/step3/strategy/RandomMoveStrategy.java
@@ -1,2 +1,13 @@
-package step3.strategy;public class RandomMoveStrategy {
+package step3.strategy;
+
+import step3.validite.RandomNumber;
+
+public class RandomMoveStrategy implements MoveStrategy {
+
+    private static final RandomNumber RANDOM_NUMBER = new RandomNumber();
+
+    @Override
+    public int move() {
+        return RANDOM_NUMBER.randomNumberCreate();
+    }
 }
